moving models to separate subpackage
